Module Superbol_free_lib.Arg_utils

val switch : ?descr:string -> [ `enable_disable | `with_without | `boolean | `use | `force ] -> name:string -> default:bool -> bool Stdlib.ref * Ezcmd.V2.EZCMD.TYPES.arg_list

switch kind ~name ~default returns a Boolean flag flg and specifications of arguments for setting/unsetting flg. The flag has default value default, and the phrasing of the documentation for arguments depends on the kind of switch: `enable_disable and `with_without are self-explanatory; `boolean uses "Set" and "Clear" verbs. `use generates a single argument selected based on the value of default.